Refactored upsert mock
